Fixed Error in Second Handling

+++ include/clockWork.h
@@ -4,7 +4,7 @@
 
 class ClockWork {
 private:
-    uint8_t countMillis48 = 0;
+    uint16_t countMillis48 = 0;
     uint32_t previous48 = 0;
     const uint16_t interval48 = 1250; // 60000/48
     uint8_t lastSecond48 = 48;
